Меняет и добавляет модули для фото, комментов и констант, а так же удаляет ненужное
