--- js/generator-id/create-picture.js
@@ -1,24 +1,0 @@
-import {DESCRIPTIONS, LIKE_MIN_COUNT, LIKE_MAX_COUNT, COMMENT_COUNT, PICTURE_COUNT} from './data.js';
-import {getRandomInteger, getRandomArrayElements} from './util.js';
-import {createComment} from './generator-id/comment-message.js';
-
-
-const createPicture = (index) => ({
-  id: index,
-  url: `photos/${index}.jpg`,
-  description: getRandomArrayElements(DESCRIPTIONS),
-  likes: getRandomInteger(LIKE_MIN_COUNT, LIKE_MAX_COUNT),
-  comment: Array.from(
-    { length: getRandomInteger(0, COMMENT_COUNT) },
-    createComment
-  ),
-});
-
-const getPictures = () =>
-  Array.from({ length: PICTURE_COUNT }, (_, pictureIndex) =>
-    createPicture(pictureIndex + 1)
-  );
-
-getPictures();// Этот вызов уместно так оставлять в этом модуле или же есть другой способ?
-
-export { createPicture, getPictures };
