+++ js/main.js
@@ -1,17 +1,3 @@
-import './util.js';
-import './data.js';
-// import './functions.js';
-
-import {createPicture, getPictures} from './generator-id/create-picture.js';
-getPictures();
-createPicture();
-
-import {createMessage, createComment} from './generator-id/comment-message.js';
-createMessage();
-createComment();
-
-import {createIdGenerator, generateCommentId} from './generator-id/create-generator-index.js';
-createIdGenerator();
-generateCommentId();
+import { getPictures } from './mocks/photo.js';
 
 getPictures();
