+++ js/mocks/comment.js
@@ -1,0 +1,27 @@
+import { getRandomArrayElements, getRandomInteger } from './util.js';
+import { COMMENTS_STROKES, AVATAR_COUNT, NAMES } from './const.js';
+
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+
+  return () => {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+const generateCommentId = createIdGenerator();
+
+const createMessage = () =>
+  Array.from({ length: getRandomInteger(1, 2) }, () =>
+    getRandomArrayElements(COMMENTS_STROKES)
+  ).join(' ');
+
+const createComment = () => ({
+  id: generateCommentId(),
+  avatar: `img/avatar-${getRandomInteger(1, AVATAR_COUNT)}.svg`,
+  message: createMessage(),
+  name: getRandomArrayElements(NAMES),
+});
+
+export { createComment };
